+++ R/mixture_functions.R
@@ -82,20 +82,40 @@ shift_pois_mix <- function(x, p, lambda, kappa) {
   return(mixture)
 }
 
-# Mixture 
-pdf_func_mix <- function(x, pars) {
+# Vectorised version of pdf_func
+pdf_func_vec <- function(pdf_func) {
+  pdf_func = match.fun(pdf_func) #solves NOTE "pdf_func is undefined"
+  
+  func <- function(x, pars){
+    n = nrow(pars)
+    output = rep(NA, n)
+    
+    for (i in 1:n) {
+      output[i] = pdf_func(x, pars[i, ])
+    }
+    
+    return(output)
+  }
+  
+  return(func)
+}
+
+# Mixture of pdf_func
+pdf_func_mix <- function(x, pars, pdf_func) {
+  pdf_func = match.fun(pdf_func) #solves NOTE "pdf_func is undefined"
+  
   mixture = 0
   for (i in 1:nrow(pars)) {
-    mixture = mixture + pars[i, ] * pdf_func(x, pars[i, -1])
+    mixture = mixture + pars[i, 1] * pdf_func(x, pars[i, -1, drop = F])
   }
   return(mixture)
 }
 
 # wrapper
 dist_mixture <- function(x, dist, pars, pdf_func = NULL) {
- 
   if (!is.null(pdf_func)) {
-    output = pdf_func_mix(x, pars)
+    output = pdf_func_mix(x, pars, pdf_func)
+    
   } else {
     if (dist == "normal") {
       output = normal_mix(x, pars[, "theta"], pars[, "mu"], pars[, "sigma"])
@@ -123,11 +143,9 @@ dist_pdf <- function(x, dist, pars, pdf_func = NULL) {
   
   if (!is.null(pdf_func)) {
     pdf_func = match.fun(pdf_func) #solves NOTE "pdf_func is undefined"
-    
-    dist_pdf = pdf_func(x, pars)
+    output = pdf_func(x, pars)
     
   } else {
-    
     if (dist == "normal") {
       output = dnorm(x, pars[, "mu"], pars[, "sigma"])
     }